+++ lib/defdo/cloudflare/monitor.ex
@@ -53,13 +53,12 @@ defmodule Defdo.Cloudflare.Monitor do
   defp execute_monitor do
     Logger.info("Executing checkup...")
 
-    get_cloudflare_config_domains()
+    get_all_cloudflare_config_domains()
     |> Enum.map(&process/1)
   end
 
   defp process(domain) do
     Logger.info("Processing domain: #{domain}")
-    local_ip = get_current_ip()
     zone_id = get_zone_id(domain)
 
     if is_nil(zone_id) do
@@ -67,14 +66,53 @@ defmodule Defdo.Cloudflare.Monitor do
       Logger.error(message)
       [message]
     else
+      a_records_to_monitor =
+        if domain_configured?(domain, :domain_mappings) do
+          records_to_monitor(domain, :domain_mappings)
+        else
+          []
+        end
+
+      aaaa_records_to_monitor =
+        if domain_configured?(domain, :aaaa_domain_mappings) do
+          records_to_monitor(domain, :aaaa_domain_mappings)
+        else
+          []
+        end
+
+      a_record_name_set = MapSet.new(a_records_to_monitor)
+      aaaa_record_name_set = MapSet.new(aaaa_records_to_monitor)
+
+      local_ipv4 =
+        if MapSet.size(a_record_name_set) > 0 do
+          get_current_ipv4()
+        else
+          nil
+        end
+
+      local_ipv6 =
+        if MapSet.size(aaaa_record_name_set) > 0 do
+          get_current_ipv6()
+        else
+          nil
+        end
+
+      if MapSet.size(a_record_name_set) > 0 and is_nil(local_ipv4) do
+        Logger.error("Unable to detect public IPv4 address; A records cannot be synchronized")
+      end
+
+      if MapSet.size(aaaa_record_name_set) > 0 and is_nil(local_ipv6) do
+        Logger.warning(
+          "Unable to detect public IPv6 address; AAAA records will be skipped for this cycle"
+        )
+      end
+
       configured_cname_records = get_cname_records_for_domain(domain)
       cname_record_names = configured_cname_records |> Enum.map(& &1["name"]) |> MapSet.new()
 
-      # obtained by Application config
-      # retrieves the subdomains to be monitored
       dns_records_to_monitor =
-        domain
-        |> records_to_monitor()
+        a_records_to_monitor
+        |> Kernel.++(aaaa_records_to_monitor)
         |> Kernel.++(MapSet.to_list(cname_record_names))
         |> Enum.uniq()
 
@@ -97,29 +135,14 @@ defmodule Defdo.Cloudflare.Monitor do
                 []
 
               get_cloudflare_key(:auto_create_missing_records) ->
-                Logger.info("Creating missing DNS record: #{record_name}")
-
-                proxied = get_cloudflare_key(:proxy_a_records, false)
-                ttl = if proxied, do: 1, else: 300
-
-                record_data = %{
-                  "type" => "A",
-                  "name" => record_name,
-                  "content" => local_ip,
-                  "ttl" => ttl,
-                  "proxied" => proxied
-                }
-
-                case create_dns_record(zone_id, record_data) do
-                  {true, result} ->
-                    Logger.info("Created DNS record: #{record_name} with promotional comment")
-
-                    [result]
-
-                  {false, _} ->
-                    Logger.error("Failed to create DNS record: #{record_name}")
-                    []
-                end
+                create_missing_ip_records(
+                  zone_id,
+                  record_name,
+                  local_ipv4,
+                  local_ipv6,
+                  a_record_name_set,
+                  aaaa_record_name_set
+                )
 
               true ->
                 Logger.info("Set AUTO_CREATE_DNS_RECORDS=true to auto-create missing records")
@@ -136,7 +159,7 @@ defmodule Defdo.Cloudflare.Monitor do
 
       ip_result =
         ip_dns_records
-        |> input_for_update_dns_records(local_ip)
+        |> input_for_update_dns_records(%{"A" => local_ipv4, "AAAA" => local_ipv6})
         |> Enum.map(fn input ->
           {success, result} = apply_update(zone_id, input)
 
@@ -185,6 +208,70 @@ defmodule Defdo.Cloudflare.Monitor do
       Logger.info("Checkup completed")
 
       result
+    end
+  end
+
+  defp create_missing_ip_records(
+         zone_id,
+         record_name,
+         local_ipv4,
+         local_ipv6,
+         a_record_name_set,
+         aaaa_record_name_set
+       ) do
+    proxied = get_cloudflare_key(:proxy_a_records, false)
+    ttl = if proxied, do: 1, else: 300
+
+    record_types_to_create =
+      []
+      |> maybe_add_missing_record_type("A", record_name, a_record_name_set, local_ipv4)
+      |> maybe_add_missing_record_type("AAAA", record_name, aaaa_record_name_set, local_ipv6)
+
+    Enum.flat_map(record_types_to_create, fn {record_type, record_ip} ->
+      Logger.info("Creating missing DNS record: #{record_type} #{record_name}")
+
+      record_data = %{
+        "type" => record_type,
+        "name" => record_name,
+        "content" => record_ip,
+        "ttl" => ttl,
+        "proxied" => proxied
+      }
+
+      case create_dns_record(zone_id, record_data) do
+        {true, result} ->
+          Logger.info(
+            "Created DNS record: #{record_type} #{record_name} with promotional comment"
+          )
+
+          [result]
+
+        {false, _} ->
+          Logger.error("Failed to create DNS record: #{record_type} #{record_name}")
+          []
+      end
+    end)
+  end
+
+  defp maybe_add_missing_record_type(
+         acc,
+         record_type,
+         record_name,
+         monitored_names,
+         detected_ip
+       ) do
+    if MapSet.member?(monitored_names, record_name) do
+      if is_binary(detected_ip) and detected_ip != "" do
+        [{record_type, detected_ip} | acc]
+      else
+        Logger.warning(
+          "Skipping #{record_type} auto-create for #{record_name}: no detected #{record_type} public address"
+        )
+
+        acc
+      end
+    else
+      acc
     end
   end